refactor: Moving some functions in a lib folder
