+++ src/lib/getCategory.js
@@ -1,3 +1,5 @@
-export const getCategory = doc => {
+const getCategory = doc => {
   return doc.manualCategoryId || doc.localCategoryId || doc.cozyCategoryId
 }
+
+module.exports = getCategory
