+++ src/lib/splitClasses.spec.js
